ecommerce-mcp-server/requirements.txt
mcp_use
pandas
langchain_google_genai
uvicorn